Combine merchant and customer ClosingBalances struct into one definition
