+++ integration_tests/tests/close.rs
@@ -6,7 +6,8 @@ use crate::{
     },
 };
 use zeekoe::{
-    customer::database::{ClosingBalances, StateName as CustomerStatus},
+    customer::database::{StateName as CustomerStatus},
+    database::ClosingBalances,
     protocol::ChannelStatus as MerchantStatus,
 };
 use zkabacus_crypto::{CustomerBalance, MerchantBalance};
